Game 2 updated
Changed the function that updates the player's HP so that it can both increase
and decrease it.
HP is now lost every time a word gets to the bottom of the screen and some
is recovered upon destroying a word.

Added an energy bar on the upper left corner that charges as words are
destroyed (on upcoming updates charging it will trigger special powers).

Added new display statistics and the corresponding locale text.

Related to #13