Game 2 updated
Implemented client-server communication with socketio (as of this update, the
server side of the communication is a placeholder and still needs to connect to
the database).

Changed the calculations for fireball movement. Fireballs now move at a fixed
speed towards their target, instead of moving at different speeds in order to take
a fixed amount of time to reach the target.

Related to #13