Games updated with graphic assets
-General:
Added new sprites and images for both game1 and game2.
Adapted some code to the new assets.
Added a logo to the loading screen for the games and centered said screen's
loading bar.
-Game2:
Fireballs are now an animated sprite instead of a plain image.
The power activation buttons are now gray when the player doesn't have
enough energy to activate them. They are colored again when the player
reaches the necesary amount of energy.